feat(web/chat): thinking shows a live peek line instead of auto-expanding
Extract the bg-task live line into a shared LivePeekLine (latest line +
throttle + fade). Thinking now stays collapsed by default and, while
streaming, peeks its latest reasoning line through the same component, so
reasoning and background output read as one system. Full reasoning is one
click away; no tall auto-expanded block thrashing the layout.
